Enable DWARF unwinders for SPU
This patch enables use of DWARF unwinders for the SPU target.

In addition to appending the DWARF unwinders, we also need to install
a spu_dwarf_reg_to_regnum that maps the raw stack pointer register to
the cooked version (to avoid mismatches with gdbarch_sp_regnum).

This also causes confusion with the AX collect handling, so we also
install ax_pseudo_register routines to handle the cooked SP.

gdb/
2014-04-17  Ulrich Weigand  <uweigand@de.ibm.com>

	* spu-tdep.c: Include "dwarf2-frame.h" and "ax.h".
	(spu_ax_pseudo_register_collect): New function.
	(spu_ax_pseudo_register_push_stack): Likewise.
	(spu_dwarf_reg_to_regnum): Likewise.
	(spu_gdbarch_init): Install them.  Append DWARF unwinders.